% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.r
\name{appsection}
\alias{appsection}
\title{Issue LaTeX section and/or subsection in appendix}
\usage{
appsection(section = NULL, subsection = NULL, main = FALSE, panel = "")
}
\arguments{
\item{section}{a character string that will cause a section command to be added to app.tex}

\item{subsection}{a character string that will cause a subsection command to be added to app.tex}

\item{main}{set to \code{TRUE} to also write a section or subsection command to the console to be used in building the main report body (graphical section), in which case you should also specify \code{panel} if option \code{texdir} is not an empty string}

\item{panel}{panel string; must be given if \code{main=TRUE} and option \code{texdir} is not \code{""}}
}
\description{
This is useful for copying section and subsection titles in the main body of the report to the appendix, to help in navigating supporting tables.  LaTeX backslash characters need to be doubled.
}

