% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.r
\name{getgreportOption}
\alias{getgreportOption}
\title{Get greport Options}
\usage{
getgreportOption(opts = NULL)
}
\arguments{
\item{opts}{character vector containing list of option names to retrieve.  If only one element, the result is a scalar, otherwise a list.  If \code{opts} is not specified, a list with all current option settings is returned.}
}
\description{
Get greport options, assigning default values of unspecified optios.
}

