% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.r
\name{latticeInit}
\alias{latticeInit}
\title{Setup lattice plots using greport options}
\usage{
latticeInit()
}
\description{
Initializes colors and other graphical attributes based on
what is stored in system option \code{greport}.
}

