% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.r
\name{maskDframe}
\alias{maskDframe}
\title{Mask Variables in a Data Frame}
\usage{
maskDframe(x, formula, ...)
}
\arguments{
\item{x}{an input data frame or data table}

\item{formula}{a formula specifying the variables to perturb}

\item{\dots}{parameters to pass to \code{maskVal}}
}
\description{
Given a list of applicable variable names in a formula, runs \code{maskVal} on any variables in a data frame \code{x} whose name is found in \code{formula}.
}

