% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.r
\name{setgreportOption}
\alias{setgreportOption}
\title{Set greport Options}
\usage{
setgreportOption(...)
}
\arguments{
\item{\dots}{a series of options for which non-default values are desired:
\itemize{
 \item{\code{tx.pch}:}{symbols corresponding to treatments}
 \item{\code{tx.col}:}{colors corresponding to treatments}
 \item{\code{tx.linecol}:}{colors for lines in line plots}
 \item{\code{nontx.col}:}{colors for categories other than treatments}
 \item{\code{tx.lty}:}{line types corresponding to treatments}
 \item{\code{tx.lwd}:}{line widths corresponding to treatments}
 \item{\code{tx.var}:}{character string name of treatment variable}
 \item{\code{er.col}:}{2-vector with names \code{"enrolled","randomized"} containing colors to use for enrolled and randomized in needle displays}
 \item{\code{alpha.f}:}{single numeric specifying alpha adjustment to be applied to all colors.  Default is 0.7}
 \item{\code{denom}:}{named vector with overall sample sizes}
 \item{\code{tablelink}:}{a character string, either \code{"tooltip"} or \code{"hyperref"} (the default); use the latter to make supporting data tables be hyperlinked to tables in the appendix rather than using a pop-up tooltip}
\item{\code{figenv}:}{LaTeX figure environment to use, default is \code{"figure"}.  Use \code{figenv="SCfigure"} if using the LaTeX \code{sidecap} package.  \code{SCfigure} is only used for narrow images, by calling \code{putFig} with \code{sidecap=TRUE}.}
\item{code{figpos}:}{LaTeX figure environment position; default is \code{"htb!"}}
 \item{\code{gtype}:}{graphics type, \code{"pdf"} or \code{"interactive"}}
 \item{\code{pdfdir}:}{name of subdirectory in which to write \code{pdf} graphics}
 \item{\code{texdir}:}{name of subdirectory in which to write \code{LaTeX} code}
 \item{\code{texwhere}:}{default is \code{"texdir"} to use location specified by \code{texdir}.  Set to \code{""} to write generated non-appendix LaTeX code to the console as expected by \code{knitr}}
\item{\code{defs}"}{fully qualified file name to which to write LaTeX macro definitions such as poptables}
}}
}
\description{
Set greport Options
}

