\name{upFirst}
\alias{upFirst}
\title{Change First Letters to Upper Case}
\usage{
upFirst(txt)
}
\arguments{
  \item{txt}{a character vector}
}
\description{
Changes the first letter of each word in a string to upper
case, keeping selected words in lower case
}
\examples{
upFirst(c('this and that','that is Beyond question'))
}
\references{
\url{http://lanecc.libguides.com/content.php?pid=38483&sid=295540},
\url{http://en.wikipedia.org/wiki/Letter_case#Headings_and_publication_titles},
\url{http://titlecapitalization.com}
}

