% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greport-package.r
\docType{package}
\name{greport}
\alias{greport}
\alias{.noGenerics}
\alias{package-greport}
\title{Graphical Reporting for Clinical Trials}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.noGenerics
}
\description{
Graphical clinical trial reporting based on Rmarkdown, LaTeX, and pdf
}
\author{
Frank E Harrell Jr \email{fh@fharrell.com}
}
\keyword{datasets}
