\docType{package}
\name{greport}
\alias{greport}
\alias{greport-package}
\alias{package-greport}
\title{Graphical Reporting for Clinical Trials}
\format{\preformatted{ logi TRUE
}}
\usage{
.noGenerics
}
\description{
Graphical Reporting for Clinical Trials
}
\author{
Frank E Harrell Jr \email{f.harrell@vanderbilt.edu}
}
\keyword{datasets}

