% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greport-package.r
\docType{package}
\name{greport}
\alias{.noGenerics}
\alias{greport}
\alias{greport-package}
\alias{package-greport}
\title{Graphical Reporting for Clinical Trials}
\format{An object of class \code{logical} of length 1.}
\usage{
.noGenerics
}
\description{
Graphical Reporting for Clinical Trials
}
\author{
Frank E Harrell Jr \email{f.harrell@vanderbilt.edu}
}
\keyword{datasets}

